+++ features/step_definitions/common_steps.rb
@@ -17,3 +17,10 @@ Given /^created old howitzer project based on cucumber$/ do
   overwrite_file('Gemfile', "Hello")
   remove_file 'config/default.yml'
 end
+
+Given /^created old howitzer project based on turnip$/ do
+  clean_current_dir
+  run_simple 'howitzer new test_automation --turnip'
+  FileUtils.move(Dir.glob("#{Dir.pwd}/#{current_dir}/test_automation/*"), File.join(Dir.pwd, current_dir))
+  FileUtils.remove_dir File.join(current_dir, 'test_automation'), true
+end